Move wagtail.widget_adapters to wagtail.admin.telepath.widgets
