+++ docs/reference/streamfield/widget_api.md
@@ -9,12 +9,12 @@ This implementation can be driven by [Stimulus](extending_client_side_stimulus) 
 The [telepath](https://wagtail.github.io/telepath/) library is used to set up mappings between Python widget classes and their corresponding JavaScript implementations. To create a mapping, define a subclass of `wagtail.widget_adapters.WidgetAdapter` and register it with `wagtail.admin.telepath.register`.
 
 ```{versionchanged} 7.1
-The `register` function should now be imported from `wagtail.admin.telepath` rather than `wagtail.telepath`.
+The `register` function should now be imported from `wagtail.admin.telepath` rather than `wagtail.telepath`, and `WidgetAdapter` should be imported from `wagtail.admin.telepath.widgets` rather than `wagtail.widget_adapters`.
 ```
 
 ```python
 from wagtail.admin.telepath import register
-from wagtail.widget_adapters import WidgetAdapter
+from wagtail.admin.telepath.widgets import WidgetAdapter
 
 class FancyInputAdapter(WidgetAdapter):
     # Identifier matching the one registered on the client side
