+++ docs/releases/7.1.md
@@ -179,9 +179,9 @@ The `AccessibilityItem` user bar item now accepts an {attr}`~wagtail.admin.userb
 
 If you do not update your code, the `AccessibilityItem` will continue to work, but the `in_editor` argument will always be `False`. Wagtail does not currently use this argument, but it may do so in the future. It may also be utilized by users for [cross-domain headless setups](headless_accessibility_checker).
 
-### `wagtail.telepath` moved to `wagtail.admin.telepath`
+### `wagtail.telepath` and `wagtail.widget_adapters` moved to `wagtail.admin.telepath`
 
-The `wagtail.telepath` module has now been moved to `wagtail.admin.telepath`. Any import lines referencing the old location need to be updated.
+The `wagtail.telepath` module has now been moved to `wagtail.admin.telepath`, and `wagtail.widget_adapters` has been moved to `wagtail.admin.telepath.widgets`. Any import lines referencing the old locations need to be updated.
 
 ## Upgrade considerations - changes to undocumented internals
