+++ wagtail/admin/rich_text/editors/draftail/__init__.py
@@ -8,8 +8,8 @@ from django.utils.functional import cached_property
 from wagtail.admin.rich_text.converters.contentstate import ContentstateConverter
 from wagtail.admin.staticfiles import versioned_static
 from wagtail.admin.telepath import register
+from wagtail.admin.telepath.widgets import WidgetAdapter
 from wagtail.rich_text import features as feature_registry
-from wagtail.widget_adapters import WidgetAdapter
 
 
 class DraftailRichTextArea(widgets.HiddenInput):
