+++ wagtail/contrib/table_block/blocks.py
@@ -11,8 +11,8 @@ from django.utils.translation import gettext as _
 
 from wagtail.admin.staticfiles import versioned_static
 from wagtail.admin.telepath import register
+from wagtail.admin.telepath.widgets import WidgetAdapter
 from wagtail.blocks import FieldBlock
-from wagtail.widget_adapters import WidgetAdapter
 
 DEFAULT_TABLE_OPTIONS = {
     "minSpareRows": 0,
